feat: move the implementation to the scheduler instead
